pyproject.toml
[project]
name = "agents-tracie"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Paiva",email = "daniel@nuvemlabs.co.nz"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "openai (>=1.82.1,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "chromadb (>=1.0.12,<2.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "google-search-results (>=2.4.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "agent", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
coverage = "^7.8.2"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.16.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
